+++ examples/cli/src/domain/mod.rs
@@ -1,0 +1,2 @@
+mod repo;
+pub use repo::*;