+++ examples/cli/src/infra/transaction.rs
@@ -1,0 +1,3 @@
+// Fake DB transaction
+#[dill::component]
+pub struct Transaction;